+++ back-end/src/main/java/site/yacht/backend/domain/deployment/service/DeployService.java
@@ -13,14 +13,11 @@ import site.yacht.backend.domain.application.exception.ApplicationNotFoundExcept
 import site.yacht.backend.domain.application.repository.ApplicationRepository;
 import site.yacht.backend.domain.deployment.dto.DeployApplicationRequestToCloud;
 
-import javax.net.ssl.*;
+import javax.net.ssl.HttpsURLConnection;
 import java.io.BufferedOutputStream;
 import java.io.OutputStream;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
-import java.security.cert.X509Certificate;
 
 @Slf4j
 @Service
@@ -43,9 +40,7 @@ public class DeployService {
         DeployApplicationRequestToCloud requestToCloud = new DeployApplicationRequestToCloud(application);
         String jsonString = "";
         try {
-            // 객체를 JSON 문자열로 변환
             jsonString = objectMapper.writeValueAsString(requestToCloud);
-            System.out.println("JSON 형식으로 변환된 데이터: " + jsonString);
         } catch (JsonProcessingException e) {
             return false;
         }
@@ -54,9 +49,9 @@ public class DeployService {
     }
 
     private int deployToCloud(String body) {
-        disableSslVerification();
         try {
             URL url = new URL(cloudUrl);
+
             HttpsURLConnection connection = (HttpsURLConnection) url.openConnection();
             connection.setRequestMethod("POST");
             connection.setRequestProperty("Content-Type", "application/json; utf-8");
@@ -71,43 +66,6 @@ public class DeployService {
         } catch (Exception e) {
             e.printStackTrace();
             throw new RuntimeException("POST request failed with exception: " + e.getMessage());
-        }
-    }
-
-    private void disableSslVerification() {
-        try {
-            // Create a trust manager that does not validate certificate chains
-            TrustManager[] trustAllCerts = new TrustManager[]{new X509TrustManager() {
-                public java.security.cert.X509Certificate[] getAcceptedIssuers() {
-                    return null;
-                }
-
-                public void checkClientTrusted(X509Certificate[] certs, String authType) {
-                }
-
-                public void checkServerTrusted(X509Certificate[] certs, String authType) {
-                }
-            }
-            };
-
-            // Install the all-trusting trust manager
-            SSLContext sc = SSLContext.getInstance("SSL");
-            sc.init(null, trustAllCerts, new java.security.SecureRandom());
-            HttpsURLConnection.setDefaultSSLSocketFactory(sc.getSocketFactory());
-
-            // Create all-trusting host name verifier
-            HostnameVerifier allHostsValid = new HostnameVerifier() {
-                public boolean verify(String hostname, SSLSession session) {
-                    return true;
-                }
-            };
-
-            // Install the all-trusting host verifier
-            HttpsURLConnection.setDefaultHostnameVerifier(allHostsValid);
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        } catch (KeyManagementException e) {
-            e.printStackTrace();
         }
     }
 }
